refactor: rename EmailSchedulingManager with EmailSchedulingService
